+++ RUFAS/data_validator.py
@@ -3,6 +3,19 @@ import re
 from enum import Enum
 from typing import Any, Callable, Union, Sequence
 
+from RUFAS.util import Aggregator
+
+AGGREGATION_FUNCTIONS: dict[
+    str, Callable[[list[float]], float] | Callable[[list[float]], float | None] | Callable[[list[Any]], Any | None]
+] = {
+    "average": Aggregator.average,
+    "division": Aggregator.division,
+    "product": Aggregator.product,
+    "standard_deviation": Aggregator.standard_deviation,
+    "sum": Aggregator.sum,
+    "difference": Aggregator.subtraction,
+}
+
 
 class ElementState(Enum):
     """
@@ -1675,7 +1688,7 @@ class CrossValidator:
         """
         pass
 
-    def _evaluate_expression(self, expression_block: dict[str, Any]) -> Any:
+    def _evaluate_expression(self, expression_block: dict[str, Any]) -> Any | list[Any]:
         """
         Evaluates an expression based on the provided expression block. This function also
         optionally adds to the alias pool if the "save_as" key is present in the expression block.
@@ -1689,8 +1702,62 @@ class CrossValidator:
         -------
         Any
             The result of the expression.
+
+        Notes
+        -----
+        Expression block:
+        >>> {
+        ...  "operation": "sum | difference | average | product | no_op", # optional, defaults to "no_op"
+        ...  "apply_to": "individual | group", # optional
+        ...  "ordered_variables": ["alias_0", "alias_1"],
+        ...  "save_as": "alias_2" # optional
+        ... }
         """
-        pass
+        operation = expression_block.get("operation", "no_op")
+        aggregator = AGGREGATION_FUNCTIONS.get(operation)
+
+        if operation not in AGGREGATION_FUNCTIONS or aggregator is None:
+            self._event_logs.append(
+                {
+                    "error": "Unknown Operation",
+                    "message": f"Unknown operation {operation} in cross validation rule.",
+                    "info_map": {
+                        "class": CrossValidator.__name__,
+                        "function": CrossValidator._evaluate_expression.__name__,
+                    },
+                }
+            )
+            raise ValueError(f"Unknown operation: {operation}")
+
+        ordered_variables: list[str] = expression_block.get("ordered_variables", [])
+        ordered_values: list[Any] = []
+        for alias_name in ordered_variables:
+            value = self._get_alias_value(alias_name)
+            ordered_values.append(value)
+
+        if len(ordered_values) > 1:
+            if "apply_to" not in expression_block:
+                self._event_logs.append(
+                    {
+                        "error": "Missing Apply To",
+                        "message": "Missing 'apply_to' key in expression block.",
+                        "info_map": {
+                            "class": CrossValidator.__name__,
+                            "function": CrossValidator._evaluate_expression.__name__,
+                        },
+                    }
+                )
+                raise ValueError("Missing 'apply_to' key in expression block.")
+            apply_to = expression_block["apply_to"]
+            if apply_to == "individual":
+                return ordered_values
+
+        result = aggregator(ordered_values)
+
+        if "save_as" in expression_block:
+            save_as_alise_name: str = expression_block["save_as"]
+            self._save_to_alias_pool(alias_name=save_as_alise_name, value=result)
+        return result
 
     def _evaluate_condition(self, condition_clause: dict[str, Any]) -> bool:
         """